Move all BHoM Filter Rules from oM.Elements to oM.FilterRules namespace.
In the oM.Elements namespace we can store only the BHoM object that have a direct corresponding object in Revit.
