+++ Revit_oM/Elements/ViewFilter.cs
@@ -1,5 +1,6 @@
 ﻿using BH.oM.Base;
 using BH.oM.Revit.Elements;
+using BH.oM.Revit.FilterRules;
 using System;
 using System.Collections.Generic;
 using System.Linq;
