+++ Revit_oM/FilterRules/FilterDoubleRule.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Adapters.Revit.Elements
+namespace BH.oM.Revit.FilterRules
 {
     public class FilterDoubleRule : FilterNumericValueRule
     {
