+++ Revit_oM/FilterRules/FilterLevelRule.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using BH.oM.Revit.Enums;
 
-namespace BH.oM.Adapters.Revit.Elements
+namespace BH.oM.Revit.FilterRules
 {
     public class FilterLevelRule : FilterRule
     {
