+++ Revit_oM/FilterRules/FilterRule.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Adapters.Revit.Elements
+namespace BH.oM.Revit.FilterRules
 {
     public abstract class FilterRule : BHoMObject
     {
